stream authenticated pages and trim bundles
