+++ apps/web/app/(app)/notifications/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoading } from "@/components/app/shared/RouteLoading";
+
+export default function Loading() {
+  return <RouteLoading label="notifications" rows={4} />;
+}
