+++ apps/web/app/(app)/prompts/loading.tsx
@@ -1,0 +1,5 @@
+import { RouteLoading } from "@/components/app/shared/RouteLoading";
+
+export default function Loading() {
+  return <RouteLoading label="prompts" />;
+}
