+++ apps/web/app/(app)/search/page.tsx
@@ -1,132 +1,40 @@
-"use client";
+import SearchClient, {
+  type SearchUser,
+} from "@/components/app/search/SearchClient";
+import { createClient } from "@/lib/supabase/server";
 
-import { useState, useRef, Suspense } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
-import Link from "next/link";
-import { Search } from "lucide-react";
-import type { User } from "@/types";
-import { Avatar } from "@/components/ui/Avatar";
+const PUBLIC_USER_FIELDS =
+  "id, username, display_name, bio, avatar_url, is_public";
 
-function SearchContent() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const [query, setQuery] = useState(searchParams.get("q") ?? "");
-  const [results, setResults] = useState<User[]>([]);
-  const [loading, setLoading] = useState(false);
-  const timerRef = useRef<ReturnType<typeof setTimeout>>(undefined);
+function sanitizeSearchFilter(value: string): string {
+  return value.replace(/[,()\\@]/g, "");
+}
 
-  function handleSearch(value: string) {
-    setQuery(value);
+export default async function SearchPage({
+  searchParams,
+}: {
+  searchParams: Promise<{ q?: string | string[] }>;
+}) {
+  const params = await searchParams;
+  const query = typeof params.q === "string" ? params.q : "";
+  let initialResults: SearchUser[] = [];
 
-    if (value.length < 2) {
-      setResults([]);
-      return;
-    }
+  if (query.length >= 2) {
+    const supabase = await createClient();
+    const safe = sanitizeSearchFilter(query);
+    const { data } = await supabase
+      .from("users")
+      .select(PUBLIC_USER_FIELDS)
+      .eq("is_public", true)
+      .or(
+        `username.ilike.%${safe}%,display_name.ilike.%${safe}%,github_username.ilike.%${safe}%`,
+      )
+      .limit(20);
 
-    clearTimeout(timerRef.current);
-    timerRef.current = setTimeout(async () => {
-      setLoading(true);
-      const params = new URLSearchParams({ q: value });
-      router.replace(`/search?${params.toString()}`, { scroll: false });
-
-      const res = await fetch(`/api/search?q=${encodeURIComponent(value)}&limit=20`);
-      const data = await res.json();
-      setResults(data.users ?? []);
-      setLoading(false);
-    }, 300);
+    initialResults = data ?? [];
   }
 
   return (
-    <>
-      <header className="sticky top-0 z-10 border-b border-border bg-background px-[var(--app-page-padding-x)] py-3">
-        <div className="flex items-center gap-3 rounded-[4px] border border-border px-4 py-2 focus-within:border-accent focus-within:ring-3 focus-within:ring-accent/15">
-          <Search size={18} className="shrink-0 text-muted" />
-          <input
-            type="search"
-            name="q"
-            value={query}
-            onChange={(e) => handleSearch(e.target.value)}
-            placeholder="Search by username or email"
-            aria-label="Search users"
-            className="flex-1 bg-transparent text-base outline-none placeholder:text-muted"
-          />
-        </div>
-      </header>
-
-      <div>
-        {query.length < 2 && !loading && results.length === 0 && (
-          <p className="px-[var(--app-page-padding-x)] py-8 text-center text-sm text-muted">
-            Search by username or email
-          </p>
-        )}
-        {loading && (
-          <p className="px-[var(--app-page-padding-x)] py-8 text-center text-sm text-muted">Searching&hellip;</p>
-        )}
-        {!loading && query.length >= 2 && results.length === 0 && (
-          <p className="px-[var(--app-page-padding-x)] py-8 text-center text-sm text-muted">
-            No users found for &ldquo;{query}&rdquo;
-          </p>
-        )}
-        {results.map((user) => {
-          const content = (
-            <>
-              <Avatar
-                src={user.avatar_url}
-                alt={user.username ?? user.display_name ?? ""}
-                fallback={user.display_name ?? user.username ?? "?"}
-                size="md"
-              />
-              <div className="flex-1 overflow-hidden">
-                <p className="font-medium">
-                  {user.username
-                    ? user.username
-                    : user.display_name ?? "New user"}
-                </p>
-                {user.username ? (
-                  user.bio && (
-                    <p className="truncate text-sm text-muted">{user.bio}</p>
-                  )
-                ) : (
-                  <p className="text-sm text-muted">
-                    Hasn&apos;t set up their profile yet
-                  </p>
-                )}
-              </div>
-            </>
-          );
-
-          return user.username ? (
-            <Link
-              key={user.id}
-              href={`/u/${user.username}`}
-              className="flex items-center gap-4 border-b border-border px-[var(--app-page-padding-x)] py-4 hover:bg-subtle"
-            >
-              {content}
-            </Link>
-          ) : (
-            <div
-              key={user.id}
-              className="flex items-center gap-4 border-b border-border px-[var(--app-page-padding-x)] py-4"
-            >
-              {content}
-            </div>
-          );
-        })}
-      </div>
-    </>
-  );
-}
-
-export default function SearchPage() {
-  return (
-    <Suspense
-      fallback={
-        <div className="px-[var(--app-page-padding-x)] py-8 text-center text-sm text-muted">
-          Loading&hellip;
-        </div>
-      }
-    >
-      <SearchContent />
-    </Suspense>
+    <SearchClient initialQuery={query} initialResults={initialResults} />
   );
 }
